Moved ParicipantListReceiver to controller package
